feat: get all todos

+++ src/app/api/seed/route.ts
@@ -4,14 +4,21 @@ import { faker } from "@faker-js/faker";
 
 export async function GET(request: NextRequest) {
   const searchParams = request.nextUrl.searchParams;
-  const max = Number(searchParams.get("max"));
+  const max = Number(searchParams.get("max") ?? "20");
 
-  const maximumNumberOfRows = isNaN(max) ? 20 : max;
+  if (isNaN(max)) {
+    return NextResponse.json(
+      {
+        message: "Invalid query param",
+      },
+      { status: 400 }
+    );
+  }
 
   await prisma.todo.deleteMany();
 
   await prisma.todo.createMany({
-    data: Array.from({ length: maximumNumberOfRows }).map(() => ({
+    data: Array.from({ length: max }).map(() => ({
       description: faker.lorem.lines({ min: 1, max: 3 }),
       complete: faker.datatype.boolean(),
     })),
